token add audit log
