+++ linkis-public-enhancements/linkis-script-dev/linkis-storage-script-dev-server/src/main/java/org/apache/linkis/filesystem/util/WorkspaceUtil.java
@@ -22,8 +22,6 @@ import org.apache.linkis.filesystem.exception.WorkSpaceException;
 import org.apache.linkis.filesystem.exception.WorkspaceExceptionManager;
 
 import java.io.File;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -104,26 +102,5 @@ public class WorkspaceUtil {
     if (!specialPattern.matcher(fileName).find()) {
       WorkspaceExceptionManager.createException(80028);
     }
-  }
-
-  /**
-   * @param plaintext
-   * @return
-   * @throws NoSuchAlgorithmException
-   */
-  public static String encrypt(String plaintext) throws NoSuchAlgorithmException {
-    // 使用 MD5 算法创建 MessageDigest 对象
-    MessageDigest md = MessageDigest.getInstance("MD5");
-    // 更新 MessageDigest 对象中的字节数据
-    md.update(plaintext.getBytes());
-    // 对更新后的数据计算哈希值，存储在 byte 数组中
-    byte[] digest = md.digest();
-    // 将 byte 数组转换为十六进制字符串
-    StringBuilder sb = new StringBuilder();
-    for (byte b : digest) {
-      sb.append(String.format("%02x", b & 0xff));
-    }
-    // 返回十六进制字符串
-    return sb.toString();
   }
 }
